requirements.txt
aws-cdk-lib==2.59.0
constructs>=10.0.0,<11.0.0
transformers==4.29.2
datasets==2.12.0
accelerate==0.19.0